# improve bin-packing performance