--- a/pkg/controllers/allocation/binpacking/packer.go
+++ b/pkg/controllers/allocation/binpacking/packer.go
@@ -152,8 +152,9 @@ func sortByResources(instanceTypes []cloudprovider.InstanceType) {
 func weightOf(instanceType cloudprovider.InstanceType) float64 {
 	return euclidean(
 		float64(instanceType.CPU().Value()),
-		float64(instanceType.Memory().ScaledValue(resource.Mega)), // 1 gb = 1 cpu
+		float64(instanceType.Memory().ScaledValue(resource.Giga)), // 1 gb = 1 cpu
 		float64(instanceType.NvidiaGPUs().Value())*1000,           // Heavily weigh gpus x 1000
+		float64(instanceType.AMDGPUs().Value())*1000,              // Heavily weigh gpus x1000
 		float64(instanceType.AWSNeurons().Value())*1000,           // Heavily weigh neurons x1000
 	)
 }