**Issue, if available:**
https://github.com/awslabs/karpenter/issues/625

**Description of changes:**
 - Increased performance of bin-packing. This specific optimization found a loop where a terminable case was not being checked which allowed a significant number of loop iterations to be skipped. 
   - The loop was responsible for packing pods on to a specific packable node. Once the node was full, the loop would continue unnecessarily. This change also adds another small optimization to the check to determine if the node is full by passing through the smallest pod from the list (it's already sorted when passed thru) and adds it to the currently reserved resources in order to see if anymore progress could possibly be made by continuing through the loop. 
   
- The terminable case optimization in the packing loop had a 30x improvement
- The smaller "min-pod" optimization resulted in a 50% improvement

**Benchmarks:**
   - 20,000 pods previously took 91 seconds and now takes about 3 seconds and drastically reduces memory usage. NOTE: the benchmark graph "post-optimization" was ran before the minPod optimization which reduced the 20,000 pod case from around 6 seconds to 3 seconds. The benchmarks were executed on 2 m5.large instance types. 

![graphs](https://github.com/bwagner5/karpenter/raw/images/benchmark_graphs.jpg)

**Testing:**
 - Performed A/B testing w/ 1:1 input and both gave the same output.  The specific test walked up cpu, memory, and pods to pack in increments of 100 milli-cores, 64mb, and 1 respectively.  The upper limits were 10,000 mc, 8GB, and 90 pods respectively. 
 - A/B tests on multiple node packings were also performed ad-hoc.



By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.